refactor: rrefactoring styling and comments and language

+++ src/app/shared/components/table/table.component.ts
@@ -1,9 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { TaskService } from '../../../services/task.service';
 import { Tarefa } from '../../../../Tarefa';
 import { ModalEditTaskComponent } from "../modal-edit-task/modal-edit-task.component";
+import { MatSnackBar } from '@angular/material/snack-bar';
+
 
 @Component({
   selector: 'app-table',
@@ -13,41 +15,41 @@ import { ModalEditTaskComponent } from "../modal-edit-task/modal-edit-task.compo
   styleUrls: ['./table.component.scss'],
 })
 export class TableComponent implements OnInit {
-  searchText: string = '';   // O texto da pesquisa
-  data: Tarefa[] = [];       // Lista de todas as tarefas
-  filteredTasks: Tarefa[] = []; // Tarefas filtradas
-  sortDescending: boolean = false; // Estado do checkbox
-  isEditModalVisible: boolean = false;  // Controla a visibilidade do modal de edição
-  selectedTask: Tarefa | null = null;  // Tarefa selecionada para edição
+  matSnackBar = inject(MatSnackBar);
+
+  searchText: string = '';
+  data: Tarefa[] = [];
+  filteredTasks: Tarefa[] = [];
+  sortDescending: boolean = false;
+  isEditModalVisible: boolean = false;
+  selectedTask: Tarefa | null = null;
+  isDeleteModalVisible: boolean = false;
 
   constructor(private taskService: TaskService) {}
 
   ngOnInit(): void {
     this.taskService.tasks$.subscribe((tasks) => {
       this.data = tasks;
-      this.applyFilter();  // Aplica o filtro logo após atualizar a lista de tarefas
+      this.applyFilter();
     });
 
-    // Carregar as tarefas ao inicializar
+
     this.taskService.getTasks().subscribe((tasks) => {
       this.data = tasks;
-      this.applyFilter();  // Aplica o filtro inicial
+      this.applyFilter();
     });
   }
 
-  // Método para aplicar o filtro baseado no texto de pesquisa
   applyFilter() {
     let filtered = this.data.filter(
       (item) =>
         item.tarefaNome.toLowerCase().includes(this.searchText.toLowerCase()) ||
-        (item.id && item.id.toString().includes(this.searchText))  // Inclui id, se presente
+        (item.id && item.id.toString().includes(this.searchText))
     );
 
-    // Aplica a ordenação com base no estado do checkbox
     this.filteredTasks = this.sortTasks(filtered);
   }
 
-  // Função que ordena as tarefas com base no estado do checkbox
   sortTasks(tasks: Tarefa[]): Tarefa[] {
     return tasks.sort((a, b) => {
       if (this.sortDescending) {
@@ -58,33 +60,48 @@ export class TableComponent implements OnInit {
     });
   }
 
-  // Método chamado quando o estado do checkbox muda
   toggleSort(event: Event) {
     this.sortDescending = (event.target as HTMLInputElement).checked;
     this.applyFilter();
   }
 
-  // Método para deletar uma tarefa
-  onDelete(tarefa: Tarefa) {
-    this.taskService.deleteTask(tarefa).subscribe({
-      next: () => {
-        this.taskService.refreshTasks();
-      },
-      error: (err) => {
-        console.error('Erro ao excluir tarefa:', err);
-      }
-    });
+  onDeleteConfirmation(tarefa: Tarefa) {
+    this.selectedTask = tarefa;
+    this.isDeleteModalVisible = true;
   }
 
-  // Método para abrir o modal de edição
+  deleteTask(tarefa: Tarefa | null) {
+    if (tarefa) {
+      this.taskService.deleteTask(tarefa).subscribe({
+        next: () => {
+          this.taskService.refreshTasks();
+          this.closeDeleteModal();
+          this.matSnackBar.open('Tarefa removida com sucesso', 'Ok', {
+            duration: 3000,
+            horizontalPosition: 'right',
+            verticalPosition: 'top',
+            panelClass: ['success-snackbar']
+          });
+        },
+        error: (err) => {
+          console.error('Erro ao excluir tarefa:', err);
+        }
+      });
+    }
+  }
+
+  closeDeleteModal() {
+    this.isDeleteModalVisible = false;
+    this.selectedTask = null;
+  }
+
   openEditModal(task: Tarefa): void {
-    this.selectedTask = task;  // Armazena a tarefa a ser editada
-    this.isEditModalVisible = true;  // Torna o modal visível
+    this.selectedTask = task;
+    this.isEditModalVisible = true;
   }
 
-  // Método para fechar o modal de edição
   closeEditModal(): void {
-    this.isEditModalVisible = false;  // Torna o modal invisível
-    this.selectedTask = null;  // Limpa a tarefa selecionada
+    this.isEditModalVisible = false;
+    this.selectedTask = null;
   }
 }
